core: Design modification, addition of a custom modeller
