+++ Sources/UltimatePortal/Caller.php
@@ -2,10 +2,8 @@
 
 namespace UltimatePortal;
 
-class Caller
-{
-    function __construct()
-    {
+class Caller {
+    function __construct(){
         global $sourcedir;
         require_once($sourcedir . '/UltimatePortal/CoreBase.php');
         foreach (CoreBase::filesAutoLoad as $file) {
@@ -13,23 +11,13 @@ class Caller
         }
     }
 
-    function subs(): Subs
-    {
-        return new Subs();
-    }
+    function subs(): Subs { return new Subs(); }
 
-    function load(): Load
-    {
-        return new Load();
-    }
+    function load(): Load { return new Load(); }
 
-    function subsBlock(): SubsBlocks
-    {
-        return new SubsBlocks();
-    }
+    function subsBlock(): SubsBlocks { return new SubsBlocks(); }
 
-    function subsUtils(): SubsUtils
-    {
-        return new SubsUtils();
-    }
+    function subsUtils(): SubsUtils { return new SubsUtils(); }
+
+    function ssi(): SSI { return new SSI(); }
 }
