+++ Sources/UltimatePortal/CoreBase.php
@@ -13,6 +13,8 @@ abstract Class CoreBase {
     const filesAutoLoad = [
         'Subs.php',
         'Load.php',
+        'SSI.php',
+        'Modeller.php',
         'SubsBlocks.php',
         'SubsUtils.php',
     ];
