+++ Themes/default/css/UltimatePortalCSS.css
@@ -446,3 +446,12 @@ h4.catbg span.left, h4.catbg2 span.left, h3.catbg span.left, h3.catbg2 span.left
     height: 33px !important;
     background: url(../images/theme/main_block.png) no-repeat 5px -999px !important;
 }
+
+.fw-bold {
+	font-weight: bold;
+}
+
+#display_head {
+	margin: 0px !important;
+	border-top: 1px solid #ddd;
+}
